fix: bring issue label limit back
