+++ frontend/src/views/project/CICDLayout.vue
@@ -131,6 +131,7 @@ providePlanContext({
   issue,
   rollout,
   taskRuns,
+  issueLabels: ref<string[]>(issue.value?.labels ?? []),
   ...planBaseContext,
 });
